+++ src/lib/core/dropdown-toolbar/public-api.ts
@@ -1,2 +1,0 @@
-export * from './dropdown-toolbar.component';
-export * from './dropdown-toolbar.module';