+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## v8.0.0 (Not yet published)
+
+### Changed
+
+- Updated to Angular 14 for building the library.
+
+### Removed
+
+-   Removed `<blui-dropdown-toolbar>` component.
+
 ## v7.0.3 (October 31, 2022)
 
 ### Fixed